+++ sdk/schemaregistry/schema-registry-avro/CHANGELOG.md
@@ -6,6 +6,11 @@
 - A new error type, `AvroError`, is added and is thrown in all error cases except for service calls.
 
 ### Breaking Changes
+- The `encodeMessageData` method has been renamed to `serialize`.
+- The `decodeMessageData` method has been renamed to `deserialize`.
+- The `MessageWithMetadata` interface has been renamed to `MessageContent`.
+- `MessageContent`'s `body` has been renamed to `data`.
+- `MessageAdapter`'s `consumeMessage` and `produceMessage` have been renamed to `consume` and `produce`.
 
 ### Bugs Fixed